Модуль клиенты

+++ app/Modules/Client/Services/ClientService.php
@@ -3,10 +3,12 @@
 namespace App\Modules\Client\Services;
 
 use App\Modules\Client\Http\Requests\ClientRequest;
+use App\Modules\Client\Http\Requests\ClientUpdateRequest;
 use App\Modules\Client\Http\Resources\ClientResource;
 use App\Modules\Client\Interfaces\ClientRepositoryInterface;
 use App\Modules\Client\Interfaces\ClientServiceInterface;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
+use Illuminate\Support\Facades\Auth;
 
 class ClientService implements ClientServiceInterface
 {
@@ -35,7 +37,7 @@ class ClientService implements ClientServiceInterface
         return new ClientResource($client);
     }
 
-    public function updateClient(ClientRequest $request, int $id): ClientResource
+    public function updateClient(ClientUpdateRequest $request, int $id): ClientResource
     {
         $client = $this->clientRepository->getByIdWithUser($id);
         $updatedClient = $this->clientRepository->update($client, $request->validated());
@@ -45,6 +47,9 @@ class ClientService implements ClientServiceInterface
     public function deleteClient(int $id): array
     {
         $client = $this->clientRepository->getByIdWithUser($id);
+        if ($client->user->id != Auth::user()->id) {
+            return ['message' => 'Ошибка удаления клиента'];
+        };
         $this->clientRepository->delete($client);
         return ['message' => 'Клиент успешно удален'];
     }
